Split and build source code
