+++ package.json
@@ -3,10 +3,22 @@
   "scripts": {
     "biome-check": "biome check --verbose",
     "biome-ci": "biome ci",
-    "jest": "jest --config=test/jest.config.js"
+    "jest": "jest --config=test/jest.config.js",
+    "build": "node build.js"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
+    "esbuild": "^0.25.2",
     "jest": "^29.7.0"
+  },
+  "obsidianManifest": {
+    "id": "automatic-table-of-contents",
+    "name": "Automatic Table Of Contents",
+    "version": "1.6.1",
+    "minAppVersion": "1.3.0",
+    "description": "Create a table of contents in a note, that updates itself when the note changes",
+    "author": "Johan Satgé",
+    "authorUrl": "https://github.com/johansatge",
+    "isDesktopOnly": false
   }
 }
